Add new Homebrew path and fix security concern
Homebrew switched to use `/opt/homebrew/bin` as default path now. Add this path to binary findings.

Remove the `--trust` option from the login command, since this will remove the multifactor option.

```
The 'login' subcommand will initialize a local cache and configuration folder,
then attempt to authenticate itself with the LastPass servers, using the
provided command line credentials or by interactively prompting (in the case of
multifactor or an unprovided password). The '--trust' option will cause
subsequent logins to not require multifactor authentication. If the
'--plaintext-key' option is specified, the decryption key will be saved to the
hard disk in plaintext.  Please note that use of this option is discouraged
except in limited situations, as it greatly decreases the security of data.
```